fix: dependency bullshit
