+++ __tests__/setup.ts
@@ -21,6 +21,7 @@ jest.mock('../src/remoteConfig', () => ({
       vordrWords: ['vordrwillcatchyou', 'andvordrwillhavefun'],
       vordrIps: ['192.0.2.0/24', '198.51.100.0/24', '203.0.113.0/24'],
       ignoredWorkEmailDomains: ['igored.com', 'ignored.org'],
+      rateLimitReputationThreshold: 10,
     } as typeof remoteConfig.vars,
   },
 }));
